일단수정

+++ happyhouse_final_be/src/main/java/com/ssafy/happyhouse/controller/BoardController.java
@@ -2,6 +2,8 @@ package com.ssafy.happyhouse.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,6 +22,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.ssafy.happyhouse.dto.Board;
 import com.ssafy.happyhouse.dto.Comment;
 import com.ssafy.happyhouse.service.BoardService;
+import com.ssafy.happyhouse.service.JwtServiceImpl;
 
 import io.swagger.annotations.ApiOperation;
 
@@ -35,9 +38,12 @@ public class BoardController {
 	@Autowired
 	private BoardService boardService;
 
+	@Autowired
+	private JwtServiceImpl jwtService;
+
 	@ApiOperation(value = "게시판 글목록", notes = "모든 게시글의 정보를 반환한다.", response = List.class)
 	@GetMapping
-	public ResponseEntity<List<Board>> retrieveBoard() throws Exception {
+	public ResponseEntity<List<Board>> retrieveBoard(HttpServletRequest request) throws Exception {
 		logger.debug("retrieveBoard - 호출");
 		return new ResponseEntity<List<Board>>(boardService.retrieveBoard(), HttpStatus.OK);
 	}
@@ -57,56 +63,96 @@ public class BoardController {
 
 	@ApiOperation(value = "게시판 글 등록", notes = "새로운 게시글 정보를 입력한다. 그리고 DB입력 성공여부에 따라 'success' 또는 'fail' 문자열을 반환한다.", response = String.class)
 	@PostMapping
-	public ResponseEntity<String> writeBoard(@RequestBody Board board) {
+	public ResponseEntity<String> writeBoard(@RequestBody Board board,HttpServletRequest request) {
 		logger.debug("writeBoard - 호출");
-		if (boardService.insertBoard(board)) {
-			return new ResponseEntity<String>(SUCCESS, HttpStatus.OK);
+		
+		if (jwtService.isUsable(request.getHeader("access-token"))) {
+			logger.info("사용 가능한 토큰!!!");
+			if (boardService.insertBoard(board)) {
+				return new ResponseEntity<String>(SUCCESS, HttpStatus.OK);
+			}else {
+				logger.error("게시글 등록 실패");
+			}
+		} else {
+			logger.error("사용 불가능 토큰!!!");
 		}
+
 		return new ResponseEntity<String>(FAIL, HttpStatus.NO_CONTENT);
 	}
 
 	@ApiOperation(value = "게시판 글정보 수정", notes = "글번호에 해당하는 게시글의 정보를 수정한다. 그리고 DB수정 성공여부에 따라 'success' 또는 'fail' 문자열을 반환한다.", response = String.class)
 	@PutMapping("{articleno}")
-	public ResponseEntity<String> updateBoard(@RequestBody Board board) {
+	public ResponseEntity<String> updateBoard(@RequestBody Board board, HttpServletRequest request) {
 		logger.debug("updateBoard - 호출");
 		logger.debug("" + board);
 
-		if (boardService.updateBoard(board)) {
-			return new ResponseEntity<String>(SUCCESS, HttpStatus.OK);
+		if (jwtService.isUsable(request.getHeader("access-token"))) {
+			logger.info("사용 가능한 토큰!!!");
+			if (boardService.updateBoard(board)) {
+				return new ResponseEntity<String>(SUCCESS, HttpStatus.OK);
+			}else {
+				logger.error("게시글 수정 실패");
+			}
+		} else {
+			logger.error("사용 불가능 토큰!!!");
 		}
+
 		return new ResponseEntity<String>(FAIL, HttpStatus.NO_CONTENT);
 	}
 
 	@ApiOperation(value = "게시판 글삭제", notes = "글번호에 해당하는 게시글의 정보를 삭제한다. 그리고 DB삭제 성공여부에 따라 'success' 또는 'fail' 문자열을 반환한다.", response = String.class)
 	@DeleteMapping("{articleno}")
-	public ResponseEntity<String> deleteBoard(@PathVariable int articleno) {
+	public ResponseEntity<String> deleteBoard(@PathVariable int articleno,HttpServletRequest request) {
 		logger.debug("deleteBoard - 호출");
-		if (boardService.deleteBoard(articleno)) {
-			return new ResponseEntity<String>(SUCCESS, HttpStatus.OK);
+		
+		if (jwtService.isUsable(request.getHeader("access-token"))) {
+			logger.info("사용 가능한 토큰!!!");
+			if (boardService.deleteBoard(articleno)) {
+				return new ResponseEntity<String>(SUCCESS, HttpStatus.OK);
+			}else {
+				logger.error("게시글 삭제 실패");
+			}
+		} else {
+			logger.error("사용 불가능 토큰!!!");
 		}
+		
 		return new ResponseEntity<String>(FAIL, HttpStatus.NO_CONTENT);
 	}
 
 	@ApiOperation(value = "댓글 입력", notes = "새로운 댓글 정보를 입력한다. 그리고 DB입력 성공여부에 따라 'success' 또는 'fail' 문자열을 반환한다.", response = String.class)
 	@PostMapping("/comment")
-	public ResponseEntity<String> writeComment(@RequestBody Comment comment) {
+	public ResponseEntity<String> writeComment(@RequestBody Comment comment,HttpServletRequest request) {
 		logger.debug("writeComment - 호출");
-		if (boardService.insertComment(comment)) {
-			return new ResponseEntity<String>(SUCCESS, HttpStatus.OK);
+		
+		if (jwtService.isUsable(request.getHeader("access-token"))) {
+			logger.info("사용 가능한 토큰!!!");
+			if (boardService.insertComment(comment)) {
+				return new ResponseEntity<String>(SUCCESS, HttpStatus.OK);
+			}else {
+				logger.error("댓글 등록 실패");
+			}
+		} else {
+			logger.error("사용 불가능 토큰!!!");
 		}
 		return new ResponseEntity<String>(FAIL, HttpStatus.NO_CONTENT);
 	}
 
 	@ApiOperation(value = "댓글 삭제", notes = "댓글번호에 해당하는 댓글의 정보를 삭제한다. 그리고 DB삭제 성공여부에 따라 'success' 또는 'fail' 문자열을 반환한다.", response = String.class)
 	@DeleteMapping("/comment/{commentno}")
-	public ResponseEntity<String> deleteComment(@PathVariable int commentno) {
+	public ResponseEntity<String> deleteComment(@PathVariable int commentno,HttpServletRequest request) {
 		logger.debug("deleteComment - 호출");
-		if (boardService.deleteComment(commentno)) {
-			return new ResponseEntity<String>(SUCCESS, HttpStatus.OK);
+		
+		if (jwtService.isUsable(request.getHeader("access-token"))) {
+			logger.info("사용 가능한 토큰!!!");
+			if (boardService.deleteComment(commentno)) {
+				return new ResponseEntity<String>(SUCCESS, HttpStatus.OK);
+			}else {
+				logger.error("댓글 삭제 실패");
+			}
+		} else {
+			logger.error("사용 불가능 토큰!!!");
 		}
 		return new ResponseEntity<String>(FAIL, HttpStatus.NO_CONTENT);
 	}
-	
-	//게시글, 댓글 기능 전부 헤더 토큰 확인해야하는거 아닌지????
 
 }
